+++ RyuUpdater/Program.cs
@@ -1,0 +1,42 @@
+﻿using System.Diagnostics;
+using System.IO.Compression;
+
+namespace RyuUpdater;
+
+public static class Program {
+    private static async Task Main(string[] args) {
+        if (args.Length == 0) {
+            Console.WriteLine("This is a helper application for SRMM. Please don't run manually!\nPress any key to exit...");
+            Console.ReadKey();
+
+            return;
+        }
+
+        var pid = int.Parse(args[0]);
+        var updateFile = args[1];
+        var targetDir = args[2];
+        var srmmFileName = args[3];
+
+        var tempDir = new FileInfo(updateFile).Directory!.FullName;
+
+        try {
+            if (pid != -1) {
+                await Process.GetProcessById(pid).WaitForExitAsync();
+            }
+        } catch {
+            // ignore
+        }
+
+        await Task.Delay(500);
+
+        ZipFile.ExtractToDirectory(updateFile, targetDir, overwriteFiles: true);
+        Directory.Delete(tempDir, recursive: true);
+
+        var srmmPath = Path.Combine(targetDir, srmmFileName);
+
+        Process.Start(new ProcessStartInfo {
+            FileName = srmmPath,
+            UseShellExecute = true
+        });
+    }
+}