+++ Utils/AssemblyVersion.cs
@@ -9,8 +9,18 @@ public static class AssemblyVersion {
     /// </summary>
     /// <returns>A <see cref="string"/>.</returns>
     public static string GetVersion() {
-        var version = Assembly.GetEntryAssembly()!.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion;
-        
-        return version;
+        return Assembly.GetEntryAssembly()!.GetCustomAttribute<AssemblyInformationalVersionAttribute>()?.InformationalVersion;
+    }
+
+    public static string GetBuildVersion() {
+        var fullVersion = GetVersion();
+
+        return fullVersion[..(fullVersion.IndexOf('-'))];
+    }
+
+    public static string GetBuildSuffix() {
+        var fullVersion = GetVersion();
+
+        return fullVersion[(fullVersion.IndexOf('-') + 1)..];
     }
 }